+++ httpie/cli.py
@@ -5,22 +5,25 @@ NOTE: the CLI interface may change before reaching v1.0.
 """
 from textwrap import dedent, wrap
 # noinspection PyCompatibility
-from argparse import (RawDescriptionHelpFormatter, FileType,
-                      OPTIONAL, ZERO_OR_MORE, SUPPRESS)
+from argparse import (
+    RawDescriptionHelpFormatter, FileType,
+    OPTIONAL, ZERO_OR_MORE, SUPPRESS
+)
 
 from httpie import __doc__, __version__
 from httpie.plugins.builtin import BuiltinAuthPlugin
 from httpie.plugins import plugin_manager
 from httpie.sessions import DEFAULT_SESSIONS_DIR
 from httpie.output.formatters.colors import AVAILABLE_STYLES, DEFAULT_STYLE
-from httpie.input import (HTTPieArgumentParser,
-                          AuthCredentialsArgType, KeyValueArgType,
-                          SEP_PROXY, SEP_CREDENTIALS, SEP_GROUP_ALL_ITEMS,
-                          OUT_REQ_HEAD, OUT_REQ_BODY, OUT_RESP_HEAD,
-                          OUT_RESP_BODY, OUTPUT_OPTIONS,
-                          OUTPUT_OPTIONS_DEFAULT, PRETTY_MAP,
-                          PRETTY_STDOUT_TTY_ONLY, SessionNameValidator,
-                          readable_file_arg, SSL_VERSION_ARG_MAPPING)
+from httpie.input import (
+    HTTPieArgumentParser, KeyValueArgType,
+    SEP_PROXY, SEP_GROUP_ALL_ITEMS,
+    OUT_REQ_HEAD, OUT_REQ_BODY, OUT_RESP_HEAD,
+    OUT_RESP_BODY, OUTPUT_OPTIONS,
+    OUTPUT_OPTIONS_DEFAULT, PRETTY_MAP,
+    PRETTY_STDOUT_TTY_ONLY, SessionNameValidator,
+    readable_file_arg, SSL_VERSION_ARG_MAPPING
+)
 
 
 class HTTPieHelpFormatter(RawDescriptionHelpFormatter):
@@ -414,8 +417,8 @@ sessions.add_argument(
 auth = parser.add_argument_group(title='Authentication')
 auth.add_argument(
     '--auth', '-a',
+    default=None,
     metavar='USER[:PASS]',
-    type=AuthCredentialsArgType(SEP_CREDENTIALS),
     help="""
     If only the username is provided (-a username), HTTPie will prompt
     for the password.
@@ -423,10 +426,21 @@ auth.add_argument(
     """,
 )
 
+
+class _AuthTypeLazyChoices(object):
+    # Needed for plugin testing
+
+    def __contains__(self, item):
+        return item in plugin_manager.get_auth_plugin_mapping()
+
+    def __iter__(self):
+        return iter(plugin_manager.get_auth_plugins())
+
+
 _auth_plugins = plugin_manager.get_auth_plugins()
 auth.add_argument(
     '--auth-type', '-A',
-    choices=[plugin.auth_type for plugin in _auth_plugins],
+    choices=_AuthTypeLazyChoices(),
     default=_auth_plugins[0].auth_type,
     help="""
     The authentication mechanism to be used. Defaults to "{default}".