+++ httpie/plugins/base.py
@@ -17,13 +17,36 @@ class AuthPlugin(BasePlugin):
 
     See <https://github.com/jkbrzt/httpie-ntlm> for an example auth plugin.
 
+    See also `test_auth_plugins.py`
+
     """
     # The value that should be passed to --auth-type
     # to use this auth plugin. Eg. "my-auth"
     auth_type = None
 
-    def get_auth(self, username, password):
+    # Set to `False` to make it possible to invoke this auth
+    # plugin without requiring the user to specify credentials
+    # through `--auth, -a`.
+    auth_require = True
+
+    # By default the `-a` argument is parsed for `username:password`.
+    # Set this to `False` to disable the parsing and error handling.
+    auth_parse = True
+
+    # If both `auth_parse` and `prompt_password` are set to `True`,
+    # and the value of `-a` lacks the password part,
+    # then the user will be prompted to type the password in.
+    prompt_password = True
+
+    # Will be set to the raw value of `-a` (if provided) before
+    # `get_auth()` gets called.
+    raw_auth = None
+
+    def get_auth(self, username=None, password=None):
         """
+        If `auth_parse` is set to `True`, then `username`
+        and `password` contain the parsed credentials.
+
         Return a ``requests.auth.AuthBase`` subclass instance.
 
         """